Split rtcclient in separate files for each class.
This is the first step to refactoring rtcclient.
