+++ camus/static/js/chat.js
@@ -5,7 +5,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { configureStore } from '@reduxjs/toolkit';
 import createSagaMiddleware from 'redux-saga';
-import { Manager } from './rtcclient.js';
+import { Manager } from './rtcclient/rtcclient';
 import App from './components/App.js';
 import connectionsReducer from './slices/connections';
 import devicesReducer from './slices/devices';
