+++ camus/static/js/rtcclient/rtcclient.js
@@ -1,0 +1,7 @@
+import adapter from 'webrtc-adapter';  // eslint-disable-line no-unused-vars
+
+import Manager from './Manager';
+import VideoPeer from './VideoPeer';
+import EventEmitter from './EventEmitter';
+
+export {Manager, VideoPeer, EventEmitter};
